Allow a range of logical shard IDs
Allow one Redis server to randomly represent a range of logical shard
IDs instead of locking it to just one.

There are "pure function" issues documented in the Lua script that limit
our ability to do this for now. But this is a first step in that
direction at least.